requirements.txt
pycardano>=0.12.0
blockfrost-python>=0.4.0
ipfshttpclient>=0.8.0
python-dotenv>=1.0.0
